Fix crossfade editor segfault due to wrong types
git-svn-id: svn://localhost/ardour2/branches/2.0-ongoing@2150 d708f5d6-7413-0410-9779-e7cbd77b26cf

+++ gtk2_ardour/canvas-waveview.h
@@ -56,7 +56,7 @@ struct _GnomeCanvasWaveViewCache
 {
     GnomeCanvasWaveViewCacheEntry* data;
     gint32                       allocated;
-    gint32                       data_size;
+    gint64                       data_size;
     gulong                       start;
     gulong                       end;
 };    
@@ -77,7 +77,7 @@ struct _GnomeCanvasWaveView
 	void (*peak_function)(void*,gulong,gulong,gulong,gpointer,guint32,double);
     gulong (*length_function)(void *);
     gulong (*sourcefile_length_function)(void*,double);
-    void (*gain_curve_function)(void *arg, double start, double end, float* vector, guint32 veclen);
+    void (*gain_curve_function)(void *arg, double start, double end, float* vector, gint64 veclen);
     void *gain_src;
 
     /** x-axis: samples per canvas unit. */
